rename train/adapters to train/frameworks
