+++ train/examples/verl/README.md
@@ -1,33 +1,4 @@
-# AWorld Train
-
-This module provides:
-- adapters/: reusable adapters for specific RL frameworks (e.g., VERL)
-- utils/: framework-agnostic utilities
-- examples/: runnable examples organized by framework
-
-## Layout
-
-```
-train/
-  adapters/
-    verl/
-      core/
-      utils/
-  utils/
-  examples/
-    verl/
-      agents/
-      configs/
-      datasets/
-      scripts/
-      logs/
-      outputs/
-```
-
-## Principles
-- Keep framework-specific logic in adapters.
-- Keep shared utilities in utils.
-- Keep runnable, thin wrappers and configs in examples.
+# Example: AWorld GAIA Agent + VeRL
 
 ## Installation
 1) Set up environment \
@@ -42,7 +13,7 @@ train/
     pip install verl==0.5.0
     ```
 
-## Quick Start (GAIA + VERL)
+## Quick Start
 ```bash
 cd train/examples/verl
 ```
@@ -55,7 +26,8 @@ python create_dataset.py \
   --test_size 100
 ```
 2) Edit configs under `train/examples/verl/configs/`.
-    - `agent.yaml`: Specify agent
+    - `os.environ["AGENT_TOOL_CONFIG_PATH"]`: Filepath for agent tools config
+    - `agent.yaml`: Specify training agent
     - `tool.yaml`: Tool environment configuration
 3) Run training:
 ```bash
@@ -63,6 +35,7 @@ bash scripts/run.sh
 ```
 
 ## Importing adapters
+
 ```python
-from train.adapters.verl.core.aworld_agent_loop import AworldAgentLoop
+from train.frameworks.verl.core.aworld_agent_loop import AworldAgentLoop
 ```
