+++ train/frameworks/verl/README.md
@@ -14,8 +14,9 @@ train/adapters/verl/
 
 ## Usage
 Import adapter entrypoints from your example code:
+
 ```python
-from train.adapters.verl.core.aworld_agent_loop import AworldAgentLoop
+from train.frameworks.verl.core.aworld_agent_loop import AworldAgentLoop
 ```
 Then implement your example-specific loop:
 ```python
